feat: inline RESOURCE_PROPERTY_NAME into AUTHORIZATIONS change set
Remove dedicated `add_resource_property_name_to_authorizations_table` change set and
add `RESOURCE_PROPERTY_NAME` column directly to original `create_authorization_permission_table` change set.

This is safe since RDBMS was not an official feature in 8.8 and incremental Liquibase files are not yet enforced.
Dropped the default value for the new column as it is now part of the initial table creation.