fix: don't accidentally skip processing after log truncation
This ensures that we always set the `skipProcessing` byte, regardless
of whether we should skip or not. This is important because log
truncation does not zero out invalidated data. New entries without the
`skipProcessing` byte set would allow that invalidated data to peek
through. If that data happened to contain a `1`, we'd unintentionally
skip processing of the new record.

This could occur:
1. When truncating uncommitted entries, for example on passive members
2. When replacing an existing entry because it is written again with a
higher term.

In versions <8.8, this could also occur after restoring from a backup
because the log is truncated down to the checkpoint position.