refactor: overwrite every byte of `LogAppendEntry`
With this we start setting unused bytes to zero. Because entries are
only serialized and written by the leader while followers copy entries
wholesale, it is safer to overwrite unused ranges. If we ever start to
use them again, a leader running an old version should not write garbage
data that another node running the new version might attach meaning to.