feat: include actual tenant id in pi modification record
Invlude the tenant owner of the process instance that
is modified in the Process Instance Modification event.

Previously, there was no tenantId property, and the
avalable getTenantId() method returned a hardcoded
<default> tenant value.

The background behind this implementation was that
this getter was never used, but we needed to
implement it for the purposes of multi-tenancy.