refactor: ensure new Intents and their mapping to ValueTypes need to be written only once
Previously, every time a new Intent was added, it was necessary to:
- add it to the Intent.INTENT_CLASSES list
- add it to both the Intent.fromProtocolValue methods, mapping it to the
correct ValueType
It is easy to forget to add it to one of these three places and have an
inconsistent definition of ValueType -> Intent mapping.
The mapping is now defined once (in the
Intent.computeValueTypeToIntentMap) and used everywhere.