build: avoid fatjar of debug cli
Now that this is included in the official dist build, we don't need to
build a standalone jar of it anymore. The utility of that was
questionable anyway, because it is tightly coupled to the exact version
broker version.

(cherry picked from commit 330af94a27d5fff9c583efd8a29f3e04c9fbc84d)