fix: Add `JobIntent.YIELD` to the whitelisted commands
If a cluster is under backpressure, JOB.YIELD may fail.
If that happens, then the job will be eventually TIMEOUT when
`JobTimeoutChecker` runs.
Once the timeout is fired, the job will be pushed immediately to the
workers.
If instead JOB.YIELD is published to the stream, the JOB will not be
timedout and it will be available to workers only via the polling
mechanism.

Currently, the broker does not have any information if the clients are
available to receive new jobs, so it will push job to them even if they
are not able to pick them up.
This can create a vicious cycle where the engine itself is generating
more load when it's already under pressure.
Whitelisting JOB.YIELD reduce the work the engine need to perform,
helping its recovery

(cherry picked from commit 6e6f1c0fd4e11ef55413acab5c02bab8546bfb6c)