feat: improved logging in case of mapping rule conflicts
This adds a more helpful log message for an edge-case where
a statical setup mapping rule in the Orchestration Cluster
conflicts with a mapping rule to be creates as part of the identity migration.
The particular conflict would be the same claimName and value pair but a different id.
The message provides guidance on how to resolve this conflict.