fix: use BackoffRetryStrategy in ProcessingScheduleServiceImpl
AbortableRetryStrategy makes the actor busy sping when it failed to
append the event to the log.
This is problematic, because when under load the engine is wasting CPU
cycles trying to append new events to the log.
This can slowdown even more the application, because the component that
is the current bottleneck is fighting for cpu shares with these jobs.

relates #35074