feat: remove TypedEventRegistry
A similar Map<Class<?>, ValueType> has been added to UnifiedRecordValue.
Added method `getValueType` to UnifiedRecordValue, implemented by
default by looking up the map.
Subclasses may override it with the right ValueType to avoid the map
lookup.

CommandRequestReader is now able to decode any Record, not just the
subset of commands it was using before, which might be a difference in
behaviour