feat: make GlobalListenerRecord.retries an integer instead of a string
The "retries" property contains the number of retries which should be
attempted when executing the user task listener.

For model-level listeners this is defined as a string since it is
possible to use an expression to define the number of retries, with the
expression being parsed in the UserTaskTransformer. Since global
listeners evaluation will not pass through the transformer anymore, it
will not be possible to parse an expression defined in a global
listener. Because of this, the retries property can be defined as an
integer directly.