feat: allow overwriting next key in the state and set a max key
This is not required during normal operation, but only required
in incidents where we want to manually recover the snapshot by overwriting
the key and the max key.

A max key is required so that, during replay it doesn't overwrite the new key
that we set for recovery.

A typical use case is as follows: Due to a bug or data corruption, the key in
the partition overflowed in to the next partition's key range. As a result, this
partition is dead and cannot process new commands any more. In this case, a manual
recovery using debug-cli to ovewrite the key is required to unblock the partition.
If we only set the key in the snapshot, during the replay it might overwrite it back
to the higher value thus blocking the partition again. To prevent that, we also use a
max key. During replay, the key is not updated if it is greater than the max key.

(cherry picked from commit 20eff7deb91812553ada07c8709781bc37b3995f)