feat: update the global listeners configuration at startup
A new initializer is introduced to:
- check if the configuration file contains a different global listeners
configuration than the currently stored one
- if so, generate a GLOBAL_LISTENER_BATCH.CONFIGURE command to update
the configuration

This logic is only applied in partition 1 to and then distributed to all
the other partitions in order to ensure that the same configuration is
shared among all the partitions.

The validation logic for global listeners configuration has been moved
from SystemContext to the new initializer since that's the place where
the configuration needs to be used.