feat: pin and unpin global listeners configuration to user task lifecycle events
On user task start events, the current global listeners configuration
(if available) is pinned to the user task, while it is unpinned on user
task end events.

This pinning is needed so that the same configurations is used from the
start to the end of each specific lifecycle event, without unexpected
changes in-between.

A reference to this "pinned" configuration is kept in the "intermediate
state" of the task in order to ensure its scope is limited to the
lifecycle event itself.
Because of this, the change needs to occur before the intermediate state
is created or deleted.

In addition to setting the reference on the task, the pinning logic
ensures that:
- a copy of the current global listeners configuration is available to
be referenced
- the global listeners state correctly keep track of which tasks are
actively referencing which configurations