deps: align camunda-xml-model used in optimize
This was still using an optimize specific version rather than the
version specified as version.camunda in the monorepos parent pom.

This caused issues when renovate tried to update only this optimize
specific dependency, detected by the maven enforcer plugin.

We can easily avoid this by using the same version in optimize as the
rest of the monorepo uses.