feat: add actor info to batch operation entity
We'll need to store the actor type and actor id in the secondary
storage.

The data structure is copied from AuditLogEntity which also has an actor
type and actor id. While these are the same, I've chosen to use a
separate enum for the types as it's fine to keep these separate and the
actor type enum for the audit log entity already had a specific name
including audit log.

Note that these values will be filled starting from 8.9 only. I'll add a
check using the brokerVersion of the record to ensure that this will
hold true, once we get to the handler changes.

For older batch operations, we'll simply use `null` because we no longer
know who started it.

I've considered using the username as a fallback, but it won't be used
in V2. It's a leftover from the Operate V1 backend.