feat: update BO entity with actor info
We should only fill the batch operation entity with actor info starting
from 8.9.0 onwards, as these fields are annotated as
@SinceVersion(8.9.0). We can check the broker version of the record for
this, to ensure we don't accidentally store it for a batch operation
created on 8.8 but exported after updating to 8.9.

The implementation is straightforward, with usage of SemanticVersion to
allow easily comparing the versions (this class is fast and uses
caching), so it's fine to use in an exporter.

We can simply return a record with the extracted values, or null values
so it can be directly used to update the entity.

I've added tests that shows some of the basic usage patterns:
- before 8.9
- after 8.9 with client id
- after 8.9 with username