fix: OpenSearch ISM URL
The AWS OpenSearch ISM calls are failing as the AWS varient require
a '/' prefix to the URL, where as the OSS version does not. We use the
same HTTP client, so not sure how it can even happen. I'm assuming somewhere
in the HTTP client the HOST & Endpoint URL's are concatenated to make
the API call. Eitherway, this fixed it.

This probably is one of the reasons some of our AWS OS ITs are failing too.