deps: override discontinued LZ4 dependency
Replaces the `org-lz4:lz4-java` dependency used by Elasticsearch 7.x with the latest version of `at.yawk.lz4:lz4-java`.
The original library has been discontinued.
The new one is a maintained community fork of it that fixes several security issues.
It and can be used as a drop-in replacement.