feat: add utility to retrieve global user task listeners from state and merge them with model-level ones
The user task listeners to be applied to a user task lifecycle event can
be:
- defined at model-level (stored in ExecutableUserTask)
- defined at cluster-level (stored in GlobalListenersState)

It is necessary to retrieve both kind of listeners and merge them,
keeping the correct ordering. In addition, each user task lifecycle
event "pins" a specific global listeners configuration (potentially
older than the current one); this "pinning" information is stored in the
intermediate state of the user task.