refactor: remove dependency on spring-boot-jackson
There is no need to couple this starter to spring-boot-jackson.

In the past it lead to the starter reusing a jackson mapper setup for
the spring web layer by default, we see no reason to keep that but would
rather decouple this starter from the jackson autoconfig. In case users already
declared their own mappers explicitly before, these will be picked up still,
so this should not be a breaking change for those who intended to do that.