fix: never invalidate written records
This fixes a potential bug where `pendingWrites` were already handed off
to the sequencer for writing but not yet serialized and written to disk,
but `tryExitOutOfErrorLoop` encounters an unexpected error and tries to
clear the `pendingWrites`. This could lead to an empty log entry being
written because the `SequencedBatch` already passed the non-empty check
in the sequencer but then doesn't have anything to serialize in the raft
layer.
This could lead to corrupted log entries of size 0 which can't be
handled properly.

We can avoid the clear of `pendingWrites` and `pendingResponses` because
they will be reinitialized on the next processing attempt, for example
in `errorHandlingInTransaction`.

(cherry picked from commit 5f9b2b4c59d27bbd6939ec5c97a1dcee3fe2f95d)