feat: Add Option to Select Secondary Storage for Load Tests
This depends on a camunda-platform-helm WIP PR that adds OpenSearch
as a Helm dependency and updates the configuration values.

It overrides the necessary configuration options to set up OpenSearch
containers. From the action, you can select which secondary storage
option to use when running the load test. Providing a secondary-storage
option is optional; if not specified, Elasticsearch will be used by
default. If an invalid value is provided, the workflow will terminate.