build: enforce spaces over tabs on json files
This just aligns what our .editorconfig already specifies: we use spaces
as tabs.

This was encountered in #42167 where the editorconfig resulted in using
spaces in a file that was generally using tabs, which was flagged by
copilot.

I've explicitly excluded anything generated or related to node, but I
guess this could also be enforced on package.json and tsconfig.json.