fix: Speed Up History Cleanup ITs
At the moment the ILM/ISM policies can take any time up to 5-10mins
to apply deletion to an index. This, for cleanup tests, set a very low
poll interval so that they run ofter, as a result speeding up the tests.

This refactor the current ElasticOpenSearch combined helper and create
DB specific helpers. The specific configuration is then moved to each
subclass.