feat: bash script for automated key overwrite procedure
For this procedure three scripts are created:
## detect-configs.sh
Helpful for generating the input parameters of the script as the name of
the stateful set, pvc and pods can be different across versions.
It outputs environment variable that should be exporter in familiar bash
syntax.

## recovery-procedure.sh
Contains all the steps of the procedure that should be run from the
operator.
It requires some env var to be set (suggested by the previous script)
except for the key values to set.

This script will go through all the nodes to fetch the ones that are
partecipating in the replication group of the requested partitions.
Those PVCs will be added to the job yaml that will run in k8s.

It will also generate the job yaml in
generated/recovery-job-partition-$PARTITION_ID-$date.yaml

The operator will be able to stop the procedure and manually apply the
yaml or let the script run it.

Then, it will monitor the job and wait for it to complete/fail

Finally, it will delete the job created.

(cherry picked from commit 523cdb926c333f8b753bb30d273ab6012d9b9c0c)