test: check that user task lifecycle events use a consistent task listeners configuration
When processing a user task lifecycle transition, the user task
listeners to be executed should be set at the start of the transition
and remain pinned even if the global configuration changes in the
meantime.

In particular, if:
- a lifecycle event is triggered just before a broker restart
- the global listeners configuration has been changed in the meantime
- a long-running task listener was running before the restart and it is
completed after the restart
then only the previously-configured listeners should be executed for
that lifecycle event even if the configuration changed.