refactor: rename method for clarity
Naming is hard, but IMO the findLastStartProcess was not clear because
what do we mean by last? It reads like the latest process. But actually
it can't provide the current version because the for loop starts at
current - 1.

I therefore opted for findPreviousVersionOfProcess. That a predicate is
provided to further filter the specific process to find is irrelevant to
the name of the method

(cherry picked from commit db369a4cea2e523af2c8df3e4b8d0a27487e65b0)

A conflict was resolved because the original commit also concerned
conditional events, which do not exist on this version.