refactor: rename method for clarity
Naming is hard, but IMO the findLastStartProcess was not clear because
what do we mean by last? It reads like the latest process. But actually
it can't provide the current version because the for loop starts at
current - 1.

I therefore opted for findPreviousVersionOfProcess. That a predicate is
provided to further filter the specific process to find is irrelevant to
the name of the method