fix: use threadsafe map for lastScheduledPositions
As HashMap is not threadsafe and lastScheduledPositions was being read
and written to from more than one thread.

This could produce weird behaviour (previously put values disappearing
etc).  Potentially this could lead to issues with the importer not
running for a specific record type + partition (or not doing so
consistently).

Refs: #42554