feat: add actor info to batch-operations spec
Allows all batch operation responses to provide actor type and actor id.

Note that the fields are only set for batch operations created on 8.9 or
later.

This is copied from AuditLogResponse, as we want to keep them separated
for now. If we feel it's useful to combine these into a single Actor -
define a single ActorTypeEnum for reuse then we can change that later
still.