test: check that completing a task with V2 API and retrieving it with V1 API returns all variables
When completing a task with the V1 APIs, all available variables
(runtime + draft + provided in the completion command) are saved as
snapshot variables. With V2 APIs, instead, only the variables provided
in the completion command are saved as snapshot variables.
Previously, the V1 search APIs only looked at snapshot variables when
retrieving completed tasks, causing variables to be missed if the task
was completed with V2 APIs.