fix: in V1 task search API, consider both runtime and snapshot variables for completed tasks
Previously, the V1 task search API only looked at snapshot variables
when retrieving tasks in status COMPLETED, with the assumption that a
full snapshot of the variables was created at completion time. This is
only true if the task is completed with V1 APIs, while V2 APIs only
store the variables provided in the command itself as snapshot
variables.

To solve this, the search logic is changed in order to collect both
runtime and snapshot variables when retrieving variables for completed
tasks, ensuring the snapshot ones take precendence. This ensures that
the snapshot values are used if available, but we can fallback to
runtime variables for the ones which were not saved as snapshots. Since
no snapshot is taken, it is possible that the retrieved runtime
variables do not contain their completion-time value, but this is the
same thing that happens with the newer V2 APIs and it is still better
than not returning them at all.

To implement this, the search for the related flow nodes cannot be
limited to "active" flow nodes anymore, because completed tasks are not
active by definition.