feat: add a posix SegmentAllocator using fallocate syscall
Almost all the code is taken from PR #9816 that tried this approach
around 3 years ago.
I temporarily set the posix allocator as the default, mostly for
testing.
The advantage of the posix allocator is that the write throughput is
halved, since we don't fill it with 0 when allocating it.
This has benefits as tail latencies for other disk operations are much
smaller due to queueing effects, especially on disk attached to the
network.

The downside is that each write is slightly slower, due to having to
zero the disk sector first.

(cherry picked from commit 66ac31fa25743fbd0250cd7cfcd019664de2e96d)