fix: use $ref rather than allOf
Currently $ref does not support a description, but that will soon be
fixed in OpenAPI 3.1.x. To deal with that, we currently use allOf as a
workaround as this does allow setting description. However, that's only
needed when we actually set a description. As this field doesn't specify
a custom description (it relies on the type having a description), we
have no need to use allOf.

Note that switching to allOf later would not have been a breaking
change. So it's also not necessary to set this already in an attempt to
be flexible in the future in case OpenAPI doesn't fix this problem.