feat: add ROOT_PROCESS_INSTANCE_KEY to RDMS data model
This just adds a new ROOT_PROCESS_INSTANCE_KEY column to various tables,
so we can later use that column to handle archiving/cleanup by root
process instance instead of by process instance.

At the moment this column will not get populated.  Just adding it so
we're ready to use it later.