refactor: improve backup store logging
This gets rid of the `LoggingBackupStore` wrapper class by logging at
call-sites instead.
Additionally, the log messages are simplified by moving all dynamic
content to the MDC. This makes it easier to grep for the messages and
allows us to filter logs on their structured data, for example for a
specific backup id.