fix: avoid duplicating variables in Tasklist V1 search
The VariableService#getRuntimeVariablesPerTaskId method queries runtime
variables given a list of GetVariableRequests. Each request contains a
list of variable names which should be queried. However, the method
implementation does not take into account the lists separately for each
request, instead it queries all requested variables (from all the
requests) for all tasks. The result is that variables which were not
requested for some tasks are returned anyway.

When using this method with multiple requests it is thus necessary to be
careful and make sure that only the actually-requested variables are
considered.