fix: self correct schedule on potential drift
A schedules determined execution can be in the past if the previous was too far behind, that results in an instant checkpoint. However, if a schedule is too far behind the next execution could still be in the past resulting in longer than expected delays since we are taking the absolute value of the delay.

With this change we always correct the instructions checkpoint time upon completion, so a drifted schedule will self correct.